+++ tests/test_utils.py
@@ -43,40 +43,3 @@ class ConvertToJD(TestCase):
         self.assertEqual(
             self.expected_jd, utils.convert_to_jd(self.jd_date),
             'Incorrect date for JD format')
-
-
-class CreateDataDir(TestCase):
-    """Tests for the ``convert_to_jd`` function"""
-
-    @classmethod
-    def setUpClass(cls):
-        """Create a temporary directory and add it to the environment"""
-        cls.test_dir = TemporaryDirectory()
-        cls.old_environ = environ.copy()
-        environ['SNDATA_DIR'] = cls.test_dir.name
-
-    @classmethod
-    def tearDownClass(cls):
-        cls.test_dir.cleanup()
-        del environ['SNDATA_DIR']
-        environ.update(cls.old_environ)
-
-    # Todo: Fail on Mac OS
-    def test_directories_are_created(self):
-        """Test the function creates subdirectories with the expected naming structure"""
-
-        survey_name = 'dummy_survey'
-        release_name = 'dummy_release'
-        expected_path = Path(self.test_dir.name) / survey_name / release_name
-        created_path = utils.create_data_dir(survey_name, release_name)
-
-        self.assertEqual(expected_path, created_path, 'Directory does not match expected name')
-        self.assertTrue(created_path.exists(), 'Directory does not exist')
-
-    def test_spaces_stripped_from_names(self):
-        """Test spaces are stripped before creatign the directories"""
-
-        survey_name = 'survey with spaces'
-        release_name = 'release with spaces'
-        created_path = utils.create_data_dir(survey_name, release_name)
-        self.assertFalse(' ' in str(created_path), 'Directory name has spaces')
